+++ src/main/java/baseball/participant/Computer.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.participant;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
